async wip: Get basic async write working
This only works with rtic, but allows for writes that use an irq to signal when
done.  This only currently supports writes that fit currently within the fifo,
supporting fifo draining through irqs will come later.

This requires a few things from rtic, one is a Mutex type that wraps around the
controller.  This means these async functions aren't methods but associated
functions.  The lock method will handle masking the irq handler for the duration
of the lock.  Second, these methods need a channel from rtic-sync to be able to
signal events from the interrupt handler.  Lastly, the `handle_irq` function
needs to be called by the actual IRQ handler.